+++ CHANGELOG.rst
@@ -7,18 +7,29 @@ All notable changes to ``sentinelsat`` will be listed here.
 ---------------------
 * Dropped support for Python 2.7. Now setuptools requires Python >= 3.6.
 
-Added
-~~~~~
+CLI changes
+~~~~~~~~~~~
+* ``--start`` and ``--end`` CLI arguments no longer use the default values of ``NOW-1DAY`` and ``NOW``. (#459 @valgur)
+* ``--footprints`` now expects a path for the output file to be specified instead of using the previous default ``<--path>/search_footprints.geojson``. (#462 @valgur, #407 @IpsumCapra)
+* ``--uuid``, ``--name`` and ``--query`` must now be repeated instead of passing a single comma-separated string.
+  Repeated keywords get treated as an "or" expression. (#463 @valgur)
 * Display DHuS server version with CLI flag ``--info`` (#367 @thomasyoung-audet)
 * Added searching by placenames with the CLI flag ``--location`` (#372 @thomasyoung-audet)
-* Added CLI support for ``--geometry`` input as a string (#381 @thomasyoung-audet)
+* Added CLI support for ``--geometry`` input as a WKT or GeoJSON string (#381 @thomasyoung-audet)
 * Download quicklooks directly with the CLI flag ``--quicklook`` (#361 @mackland)
+* Got rid of the special handling of ``--uuid`` and ``--name`` CLI arguments. The product IDs are now simply passed to ``api.query()`` as a set.
+  As a result they no longer ignore the date range arguments (fixes #387). (#390 @valgur)
+* Added ``--timeout`` option with a default value of 60 seconds to avoid waiting indefinitely for a response. (#475 @valgur)
+* New ``--gnss`` CLI flag to use the GNSS end-point for orbit products query and download (#441 @avalentino).
+
+Added
+~~~~~
 * Added ``setinelsat/__main__.py`` (#412 @avalentino)
 * Added ``get_stream()`` (#430 @fwfichtner)
 * New ``sentinelsat/products.py`` module providing a "product nodes" API that
   allows to filter and download only selected files of the requested products
   (#414 @avalentino)
-* New ``--gnss`` CLI flag to use the GNSS end-point for orbit products query and download (#441 @avalentino).
+* Added ``download_quicklooks()`` and ``download_all_quicklooks()`` (#361 @mackland)
 
 Changed
 ~~~~~~~
@@ -40,11 +51,15 @@ Changed
 * Fixed the server-side space handling change issue reported in #383 by quoting instead of backslash-escaping the strings. (#390 @valgur)
 * Queries now use ``(key:value OR key:value)`` format instead of the previous ``key:(value OR value)`` format,
   which for some reason took tens of seconds per query. (#390 @valgur)
-* Got rid of the special handling of ``--uuid`` and ``--name`` CLI arguments. The product IDs are now simply passed to ``api.query()`` as a set.
-  They no longer also ignore the date range arguments (fixes #387). Add ``--start *`` to CLI arguments to maintain old behavior. (#390 @valgur)
 * Empty queries raise a ``ValueError`` immediately on the client side instead of letting the server generate it. (#390 @valgur)
 * Added stricter checks for empty keyword values in queries, which would cause server-side errors. (#390 @valgur)
 * Gracefully handle cancelled futures. (#448 @avalentino)
+* Use the HTTP status instead of OData metadata to determine the online status of a product when downloading. 
+  This is a workaround for the rare server-side bug of the OData info for the online status being incorrect (#467). (#469 @valgur) 
+* Set the default query timeout to 60 seconds to avoid waiting indefinitely for a response. (#475 @valgur)
+* ``download_all`` now returns a named tuple for clarity. (#477 @valgur)
+* Search queries now use GET instead of POST requests because the Sentinel-5P Pre-Operations Data Hub does not support POST (#380).
+  As a consequence, the search query contents are now limited to latin1 encoding instead of UTF-8. (#473 @valgur)
 
 Deprecated
 ~~~~~~~~~~
@@ -52,11 +67,14 @@ Deprecated
 
 Fixed
 ~~~~~
-* fix location information for Nominatim bounding box queries (#384)
-* Get file name extension more reliably from either header or internal logic (in particular for S5 products #270) (#378 @valgur)
+* Fix location information for Nominatim bounding box queries (#384)
+* Get file name extension more reliably from either the HTTP header or an OData attribute. (#378, #472 @valgur)
 * Updated the API Hub URL to `https://apihub.copernicus.eu/apihub/`.
 * Server-side error info has become much more detailed and the client code has been updated to correctly handle that.
 * ``check_existing()`` now determines the filename correctly for Sentinel-5 products. (@valgur #452)
+* Fix accidental downloading of the whole product in memory when the product is actually available despite being marked
+  as offline in its metadata. (#386, #421, #454 @lucadelu)
+* Fixed timeout not being used in some queries. (#454 @valgur)
 
 Development Changes
 ~~~~~~~~~~~~~~~~~~~